move SpiNNaker to spynnaker.pyNN
